+++ docs/changelog.rst
@@ -7,6 +7,7 @@ Version 0.28
 Currently in development.
 
 - Fixed search via API in combination with name_only parameter
+- Fixed search for datetime attributes not being timezone-aware
 - Added ``timeline`` array style for arrays of objects with a ``datetime`` property
 - Added search to text fields with more than 10 choices (configurable with ``SAMPLEDB_MIN_NUM_TEXT_CHOICES_FOR_SEARCH``)
 - Added support for ``"first"`` and ``"last"`` statistic for timeseries